# data/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ACG Theogony API",
    "description": "ACG × 神话原型 关系知识图谱 API",
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/ai/ask": {
      "post": {
        "tags": [
          "ai"
        ],
        "summary": "Ask",
        "operationId": "ask_api_ai_ask_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AskRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AskResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/ai/nlq": {
      "post": {
        "tags": [
          "ai"
        ],
        "summary": "Nlq",
        "operationId": "nlq_api_ai_nlq_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NlqRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/ai/whatif": {
      "post": {
        "tags": [
          "ai"
        ],
        "summary": "Whatif",
        "operationId": "whatif_api_ai_whatif_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WhatIfRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WhatIfResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/ai/suggest": {
      "get": {
        "tags": [
          "ai"
        ],
        "summary": "Suggest",
        "operationId": "suggest_api_ai_suggest_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SuggestionDTO"
                  },
                  "title": "Response Suggest Api Ai Suggest Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/ai/research": {
      "post": {
        "tags": [
          "ai"
        ],
        "summary": "Research",
        "description": "gpt-luna-5.6 联网研究（需配置 LUNA_*）。",
        "operationId": "research_api_ai_research_post",
        "parameters": [
          {
            "name": "question",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Question"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AskCitation": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "relation": {
            "type": "string",
            "title": "Relation",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "title": "AskCitation"
      },
      "AskRequest": {
        "properties": {
          "question": {
            "type": "string",
            "maxLength": 300,
            "title": "Question"
          }
        },
        "type": "object",
        "required": [
          "question"
        ],
        "title": "AskRequest"
      },
      "AskResponse": {
        "properties": {
          "answer": {
            "type": "string",
            "title": "Answer"
          },
          "citations": {
            "items": {
              "$ref": "#/components/schemas/AskCitation"
            },
            "type": "array",
            "title": "Citations",
            "default": []
          },
          "subgraph": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GraphDTO"
              },
              {
                "type": "null"
              }
            ]
          },
          "engine": {
            "type": "string",
            "title": "Engine",
            "default": "graph"
          }
        },
        "type": "object",
        "required": [
          "answer"
        ],
        "title": "AskResponse"
      },
...
      "GraphDTO": {
        "properties": {
          "nodes": {
            "items": {
              "$ref": "#/components/schemas/GraphNodeDTO"
            },
            "type": "array",
            "title": "Nodes"
          },
          "links": {
            "items": {
              "$ref": "#/components/schemas/GraphLinkDTO"
            },
            "type": "array",
            "title": "Links"
          },
          "meta": {
            "additionalProperties": true,
            "type": "object",
            "title": "Meta",
            "default": {}
          }
        },
        "type": "object",
        "required": [
          "nodes",
          "links"
        ],
        "title": "GraphDTO"
      },
      "GraphLinkDTO": {
        "properties": {
          "source": {
            "type": "string",
            "title": "Source"
          },
          "target": {
            "type": "string",
            "title": "Target"
          },
          "type": {
            "$ref": "#/components/schemas/RelationshipType"
          },
          "directed": {
            "type": "boolean",
            "title": "Directed",
            "default": true
          },
          "confidence": {
            "type": "string",
            "title": "Confidence",
            "default": "verified"
          }
        },
        "type": "object",
        "required": [
          "source",
          "target",
          "type"
        ],
        "title": "GraphLinkDTO"
      },
      "GraphNodeDTO": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "kind": {
            "type": "string",
            "title": "Kind",
            "default": "character"
          },
          "mythology": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mythology"
          },
          "className": {
            "type": "string",
            "title": "Classname",
            "default": ""
          },
          "imageUrl": {
            "type": "string",
            "title": "Imageurl",
            "default": ""
          },
          "degree": {
            "type": "integer",
            "title": "Degree",
            "default": 0
          },
          "wikiId": {
            "type": "integer",
            "title": "Wikiid",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "title": "GraphNodeDTO"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "NlqRequest": {
        "properties": {
          "query": {
            "type": "string",
            "maxLength": 200,
            "title": "Query"
          }
        },
        "type": "object",
        "required": [
          "query"
        ],
        "title": "NlqRequest"
      },
...
      "RelationshipType": {
        "type": "string",
        "enum": [
          "PARENT_OF",
          "SIBLING_OF",
          "SPOUSE_OF",
          "LOVER_OF",
          "MASTER_OF",
          "ALLY_OF",
          "ENEMY_OF",
          "FOUGHT_WITH",
          "MENTOR_OF",
          "PROTOTYPE_IS",
          "GENDER_SWAP",
          "DERIVED_FROM",
          "COMPOSITE_OF",
          "BELONGS_TO"
        ],
        "title": "RelationshipType",
        "description": "关系类型（存储层只用规范方向，逆关系在入库时归一化）。"
      },
...
      "SuggestionDTO": {
        "properties": {
          "sourceId": {
            "type": "string",
            "title": "Sourceid"
          },
          "targetId": {
            "type": "string",
            "title": "Targetid"
          },
          "commonNeighbors": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Commonneighbors",
            "default": []
          },
          "sameMythology": {
            "type": "boolean",
            "title": "Samemythology",
            "default": false
          },
          "score": {
            "type": "number",
            "title": "Score",
            "default": 0.0
          }
        },
        "type": "object",
        "required": [
          "sourceId",
          "targetId"
        ],
        "title": "SuggestionDTO"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WhatIfRequest": {
        "properties": {
          "aId": {
            "type": "string",
            "title": "Aid"
          },
          "bId": {
            "type": "string",
            "title": "Bid"
          }
        },
        "type": "object",
        "required": [
          "aId",
          "bId"
        ],
        "title": "WhatIfRequest"
      },
      "WhatIfResponse": {
        "properties": {
          "narrative": {
            "type": "string",
            "title": "Narrative"
          },
          "engine": {
            "type": "string",
            "title": "Engine",
            "default": "graph"
          }
        },
        "type": "object",
        "required": [
          "narrative"
        ],
        "title": "WhatIfResponse"
      }
    }
  }
}